finalizing project for week 1

+++ camsApp2/camsApp2/vipTickets.h
@@ -19,7 +19,7 @@ typedef enum{
     
 }vipEnum;
 
-@property int skyBoxUpgradePrice;
+@property float skyBoxUpgradePrice;
 @property int upgradeDiscount;
 @property float totalUpgradePurchase;
